updated for our environment
We will be using the shared folder in the git repo for our web root, so this needs to be setup in vagrant config. We can still set permissions, but they should already be set. We use a local config file rather than config.php or config_override.php so we will include it and copy it into place: db_config.php